requirements.txt
asgiref==3.8.1
boto3==1.37.4
botocore==1.37.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
feedparser==6.0.11
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.3
sgmllib3k==1.0.0
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
